# micro/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/role/authUser/allocatedList:
        post:
            tags:
                - SysRole
            operationId: SysRole_AllocatedList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.IsAllocatedListRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.IsAllocatedListReply'
    /v1/role/authUser/cancel:
        put:
            tags:
                - SysRole
            operationId: SysRole_AuthUserCancel
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.AuthUserCancelRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/authUser/cancelAll:
        put:
            tags:
                - SysRole
            operationId: SysRole_AuthUserCancelAll
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.AuthUserSelectAllRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/authUser/selectAll:
        put:
            tags:
                - SysRole
            operationId: SysRole_AuthUserSelectAll
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.AuthUserSelectAllRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/authUser/unallocatedList:
        post:
            tags:
                - SysRole
            operationId: SysRole_UnAllocatedList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.IsAllocatedListRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.IsAllocatedListReply'
    /v1/role/changeStatus:
        put:
            tags:
                - SysRole
            operationId: SysRole_ChangeStatusSysRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ChangeStatusSysRoleRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/dataScope:
        put:
            tags:
                - SysRole
            operationId: SysRole_DataScopeSysRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.DataScopeSysRoleRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/delete/{id}:
        delete:
            tags:
                - SysRole
            operationId: SysRole_DeleteSysRole
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/info/{roleId}:
        get:
            tags:
                - SysRole
            operationId: SysRole_GetSysRole
            parameters:
                - name: roleId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GetSysRoleReply'
    /v1/role/list:
        post:
            tags:
                - SysRole
            operationId: SysRole_ListSysRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ListSysRoleRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListSysRoleReply'
    /v1/role/save:
        post:
            tags:
                - SysRole
            operationId: SysRole_CreateSysRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ModifySysRoleRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
    /v1/role/update:
        put:
            tags:
                - SysRole
            operationId: SysRole_UpdateSysRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ModifySysRoleRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptyReply'
...
components:
    schemas:
...
        api.system.v1.AuthUserCancelRep:
            type: object
            properties:
                roleId:
                    type: string
                userId:
                    type: string
        api.system.v1.AuthUserSelectAllRep:
            type: object
            properties:
                roleId:
                    type: string
                userIds:
                    type: string
...
        api.system.v1.ChangeStatusSysRoleRep:
            type: object
            properties:
                roleId:
                    type: string
                status:
                    type: string
...
        api.system.v1.DataScopeSysRoleRep:
            type: object
            properties:
                roleId:
                    type: string
                deptIds:
                    type: array
                    items:
                        type: string
                dataScope:
                    type: string
...
        api.system.v1.DeptReply:
            type: object
            properties:
                createdAt:
                    type: string
                remark:
                    type: string
                deptId:
                    type: string
                parentId:
                    type: string
                ancestors:
                    type: string
                deptName:
                    type: string
                orderNum:
                    type: string
                leader:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                status:
                    type: string
                parentName:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.DeptReply'
            description: 部门通用返回信息
...
        api.system.v1.EmptyReply:
            type: object
            properties: {}
...
        api.system.v1.GetSysRoleReply:
            type: object
            properties:
                role:
                    $ref: '#/components/schemas/api.system.v1.RoleReply'
...
        api.system.v1.IsAllocatedListRep:
            type: object
            properties:
                pageNum:
                    type: string
                    description: pageNum
                pageSize:
                    type: string
                    description: 页面大小
                roleId:
                    type: string
        api.system.v1.IsAllocatedListReply:
            type: object
            properties:
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.UserReply'
                total:
                    type: string
...
        api.system.v1.ListSysRoleRep:
            type: object
            properties:
                pageNum:
                    type: string
                    description: pageNum
                pageSize:
                    type: string
                    description: 页面大小
        api.system.v1.ListSysRoleReply:
            type: object
            properties:
                total:
                    type: string
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.RoleReply'
...
        api.system.v1.ModifySysRoleRep:
            type: object
            properties:
                remark:
                    type: string
                roleId:
                    type: string
                roleName:
                    type: string
                roleKey:
                    type: string
                roleSort:
                    type: string
                dataScope:
                    type: string
                menuCheckStrictly:
                    type: boolean
                deptCheckStrictly:
                    type: boolean
                status:
                    type: string
                flag:
                    type: boolean
                menuIds:
                    type: array
                    items:
                        type: string
                deptIds:
                    type: array
                    items:
                        type: string
                deptNames:
                    type: array
                    items:
                        type: string
                admin:
                    type: boolean
...
        api.system.v1.RoleReply:
            type: object
            properties:
                remark:
                    type: string
                roleId:
                    type: string
                roleName:
                    type: string
                roleKey:
                    type: string
                roleSort:
                    type: string
                dataScope:
                    type: string
                menuCheckStrictly:
                    type: boolean
                deptCheckStrictly:
                    type: boolean
                status:
                    type: string
                flag:
                    type: boolean
                menuIds:
                    type: array
                    items:
                        type: string
                deptIds:
                    type: array
                    items:
                        type: string
                deptNames:
                    type: array
                    items:
                        type: string
                admin:
                    type: boolean
                createdAt:
                    type: string
            description: 角色通用返回信息
...
        api.system.v1.UserReply:
            type: object
            properties:
                remark:
                    type: string
                userId:
                    type: string
                deptId:
                    type: string
                userName:
                    type: string
                nickName:
                    type: string
                email:
                    type: string
                phoneNumber:
                    type: string
                sex:
                    type: string
                avatar:
                    type: string
                password:
                    type: string
                status:
                    type: string
                delFlag:
                    type: string
                loginIp:
                    type: string
                loginDate:
                    type: string
                dept:
                    $ref: '#/components/schemas/api.system.v1.DeptReply'
                roleIds:
                    type: array
                    items:
                        type: string
                postIds:
                    type: array
                    items:
                        type: string
                admin:
                    type: boolean
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.RoleReply'
                createdAt:
                    type: string
            description: 用户通用返回信息
...